api/requirements.txt
asgiref==3.7.2
distlib==0.3.4
Django==4.2.10
django-filter==23.5
djangorestframework==3.14.0
filelock==3.6.0
importlib-metadata==7.0.1
keyboard==0.13.5
Markdown==3.5.2
platformdirs==2.5.1
pynput==1.7.6
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
virtualenv==20.13.3
zipp==3.17.0
django-cors-headers==4.4.0
